LayoutDiffusion/requirements.txt
# The following can be installed with pip using 'pip install -r requirements.txt'
# Make sure you have upgraded pip
# pip install --upgrade pip

pillow
einops
spacy
huggingface_hub
typing-inspect
typing_extensions
scipy
pytorch_fid
seaborn
pycocotools